Conflicts:
	mrViewer/src/po/cs.po
	mrViewer/src/po/de.po
	mrViewer/src/po/es.po
	mrViewer/src/po/fr.po
	mrViewer/src/po/it.po
	mrViewer/src/po/ja.po
	mrViewer/src/po/ko.po
	mrViewer/src/po/messages.pot
	mrViewer/src/po/pl.po
	mrViewer/src/po/pt.po
	mrViewer/src/po/ro.po
	mrViewer/src/po/ru.po
	mrViewer/src/po/tr.po
	mrViewer/src/po/zh.po

+++ mrViewer/src/core/YouTube.cpp
@@ -48,8 +48,8 @@ namespace mrv
     std::string escape_url( const std::string& url )
     {
         char buf[64];
-        unsigned len = url.size();
-        sprintf( buf, "%%%d%%", len );
+        size_t len = url.size();
+        sprintf( buf, "%%%zd%%", len );
         return buf + url;
     }
